refactor: split two self-contained blocks out of mcp/server.py
`mureo/mcp/server.py` was 1,587 lines with thirty-three top-level
definitions spanning env gating, tool-registry construction, plugin
collection, schema validation, throttling, policy gates, result
decoration, dispatch and instructions composition. Every recent change
to it paid for that: unrelated features merge-conflicted inside one
file, and a reviewer had to re-derive which definition served which
concern.

Two clusters move out verbatim. Both were chosen for the same property:
every function in them is a function of its arguments and holds no
module state, so nothing a test monkey-patches on `server` is read from
either file.

  mureo/mcp/_plugin_declarations.py (283 lines)
    The five best-effort registrars that publish a plugin's — or a
    bridged surface's — declared budget / bid / readOnlyHint semantics
    to the policy registries, plus the bridged-table ownership check
    they share.

  mureo/mcp/_result_decorations.py (232 lines)
    The six helpers the dispatcher wraps around a tool call: the
    policy-gate refusal payload, the batch / strategy / plugin-strategy
    / learning-reset notices, and the pre-mutation plugin reversal
    capture.

server.py: 1,587 -> 1,163 lines.

What deliberately did NOT move, and why it cannot
-------------------------------------------------
The registry, gating, plugin-collection, validator, throttle, policy-gate,
dispatch and instructions layers stay in server.py. Two mechanisms pin
them there, and both are load-bearing rather than incidental:

  - `importlib.reload(mureo.mcp.server)` is how the env-gating and
    plugin-wiring suites re-trigger startup under a changed
    `MUREO_DISABLE_*` or a patched provider registry. Reload re-executes
    exactly one module body, so every module-level statement that builds
    `_ALL_TOOLS`, `_PLUGIN_*`, `_TOOL_VALIDATORS` or registers the
    declarations must remain in it. The registrar *calls* accordingly
    stayed behind; only the functions they call moved.

  - The suites monkey-patch `_policy_gate_entry_points`,
    `_load_policy_gates`, `_evaluate_policy_gates`, `_PLUGIN_THROTTLER`,
    `_throttle_store_seeded`, `_staleness_warned`, `_ALL_TOOLS`,
    `_PLUGIN_DISPATCH`, `capture_before_state`, `record_native_mutation`
    and `handle_{google,meta}_ads_tool` on the server module. A patch is
    only visible to a caller that resolves the name in the same module
    namespace, so each of those chains has to stay whole. That is what
    keeps `_plugin_tool_owners` (reads the patched `_ALL_TOOLS`) and the
    whole instructions block next to it, even though issue #678 proposed
    hoisting them.

No behaviour change, no API change
----------------------------------
Every moved name is re-exported from `mureo.mcp.server`, so
`mureo.mcp.server.X` and `from mureo.mcp.server import X` resolve
exactly as before for the fifteen callers outside the module.

Asserted mechanically rather than by inspection: all thirty-three
top-level definitions of the original file are present in the new set
with identical `ast.dump` output — the twelve that moved included — and
every module-level statement of the original survives in server.py in
the same relative order except four import statements removed
(`inspect`, `pathlib.Path`, `STOP_EXCEPTIONS`, `MCPReversibleToolProvider`,
all now unused there) and three added (the two new modules, plus the
narrowed `tool_provider` import).

Not one existing test is modified. `_declares_from_bridged_table` is the
only name not re-exported; it is a private helper of the moved block and
nothing outside it ever referenced it.

Refs #678
